% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/care-support.r
\name{CARE2_boot}
\alias{CARE2_boot}
\title{CARE Bootstrap Analysis}
\usage{
CARE2_boot(
  gamma.exp_sel,
  gamma.out_sel,
  se.exp_sel,
  se.out_sel,
  nx,
  ny,
  nrep = 1000,
  algorithm = "Lasso",
  random_start = 0,
  biascorrect = "no",
  etamean = 0.5,
  pthr = 5e-05
)
}
\arguments{
\item{gamma.exp_sel}{Vector of genetic associations with the exposure}

\item{gamma.out_sel}{Vector of genetic associations with the outcome}

\item{se.exp_sel}{Vector of standard errors for the exposure associations}

\item{se.out_sel}{Vector of standard errors for the outcome associations}

\item{nx}{Sample size for exposure}

\item{ny}{Sample size for outcome}

\item{nrep}{Number of bootstrap repetitions}

\item{algorithm}{Algorithm to use: "Lasso" or "CD" (coordinate descent)}

\item{random_start}{Number of random starting points}

\item{biascorrect}{Bias correction method: "no", "direct", "rerand", "rerand2", or "rerand3"}

\item{etamean}{Mean of the random variable for re-randomization}

\item{pthr}{P-value threshold for selecting significant SNPs}
}
\value{
A list containing CARE analysis results
}
\description{
Performs CARE analysis with bootstrapping for robust estimation.
}
\examples{
\dontrun{
results <- CARE2_boot(beta_exp, beta_out, se_exp, se_out, 10000, 10000)
}
}
