% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrcare-configuration.r
\name{check_hf_connection}
\alias{check_hf_connection}
\title{Check Hugging Face Connection}
\usage{
check_hf_connection(verbose = TRUE)
}
\arguments{
\item{verbose}{Whether to print detailed messages}
}
\value{
Logical indicating if the connection was successful
}
\description{
Tests the connection to the Hugging Face repository
}
\keyword{internal}
