% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/care-support.r
\name{correct_bias}
\alias{correct_bias}
\title{Correct Winner's Curse Bias in GWAS Summary Statistics}
\usage{
correct_bias(gamma.exp, gamma.out, se.exp, se.out, pthr)
}
\arguments{
\item{gamma.exp}{Vector of genetic associations with the exposure}

\item{gamma.out}{Vector of genetic associations with the outcome}

\item{se.exp}{Vector of standard errors for the exposure associations}

\item{se.out}{Vector of standard errors for the outcome associations}

\item{pthr}{P-value threshold for selecting significant SNPs}
}
\value{
A list with bias-corrected associations
}
\description{
Corrects bias in genetic associations due to winner's curse.
}
\keyword{internal}
