% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrcare-configuration.r
\name{get_mrcare_config}
\alias{get_mrcare_config}
\title{Get MRcare Configuration}
\usage{
get_mrcare_config(create_if_missing = TRUE, verbose = FALSE)
}
\arguments{
\item{create_if_missing}{Whether to run configuration setup if no configuration is found}

\item{verbose}{Whether to display messages during configuration}
}
\value{
A list containing configuration parameters
}
\description{
Retrieves the saved MRcare configuration or creates a new one if none exists.
}
\examples{
\dontrun{
config <- get_mrcare_config()
}
}
