% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrcare-configuration.r
\name{get_os_type}
\alias{get_os_type}
\title{Get OS Type}
\usage{
get_os_type()
}
\value{
Character string: "windows", "mac", or "linux"
}
\description{
Determines the operating system type
}
\keyword{internal}
