% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrcare-configuration.r
\name{get_plink_download_url}
\alias{get_plink_download_url}
\title{Get PLINK Download URL}
\usage{
get_plink_download_url(os_type)
}
\arguments{
\item{os_type}{Operating system type}
}
\value{
URL to download PLINK
}
\description{
Gets the download URL for PLINK based on OS
}
\keyword{internal}
