% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.r
\name{plot_funnel}
\alias{plot_funnel}
\title{Funnel Plot for MR Analysis}
\usage{
plot_funnel(
  care_results,
  highlight_invalid = TRUE,
  xlab = "MR estimate",
  ylab = "Precision (1/SE)",
  title = "MR Funnel Plot",
  colors = NULL
)
}
\arguments{
\item{care_results}{Results object from mr_care function}

\item{highlight_invalid}{Whether to highlight invalid instruments. Default is TRUE}

\item{xlab}{X-axis label. Default is "MR estimate"}

\item{ylab}{Y-axis label. Default is "Precision (1/SE)"}

\item{title}{Plot title. Default is "MR Funnel Plot"}

\item{colors}{Vector of colors for valid and invalid instruments. Default is NULL (uses preset colors)}
}
\value{
A ggplot2 object
}
\description{
Creates a funnel plot to visualize potential pleiotropy in MR analysis
}
\examples{
\dontrun{
results <- mr_care(exposure_data, outcome_data)
plot_funnel(results)
}
}
