% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{print.mrcare}
\alias{print.mrcare}
\title{Print method for summary.mrcare objects}
\usage{
\method{print}{mrcare}(x, ...)
}
\arguments{
\item{x}{A summary.mrcare object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the summary object
}
\description{
Print method for summary.mrcare objects
}
