% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/care-support.r
\name{rerandomization_bias}
\alias{rerandomization_bias}
\title{Re-randomization Bias Correction}
\usage{
rerandomization_bias(
  gamma.exp,
  gamma.out,
  se.exp,
  se.out,
  etamean = 0.5,
  pthr = 5e-05,
  var.type = "upper"
)
}
\arguments{
\item{gamma.exp}{Vector of genetic associations with the exposure}

\item{gamma.out}{Vector of genetic associations with the outcome}

\item{se.exp}{Vector of standard errors for the exposure associations}

\item{se.out}{Vector of standard errors for the outcome associations}

\item{etamean}{Mean of the random variable for re-randomization}

\item{pthr}{P-value threshold for selecting significant SNPs}

\item{var.type}{Variance type ("upper" or "org")}
}
\value{
A list with bias-corrected associations
}
\description{
Corrects bias using a re-randomization approach.
}
\keyword{internal}
