% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrcare-configuration.r
\name{setup_plink}
\alias{setup_plink}
\title{Setup PLINK}
\usage{
setup_plink(
  plink_path = NULL,
  cache_dir,
  force_download = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{plink_path}{Optional path to existing plink executable}

\item{cache_dir}{Directory to store downloaded files}

\item{force_download}{Whether to force re-download}

\item{verbose}{Whether to print detailed messages}
}
\value{
Path to plink executable
}
\description{
Checks for and installs PLINK if needed
}
\keyword{internal}
