% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{summary.mrcare}
\alias{summary.mrcare}
\title{Summarize MRcare Analysis Results}
\usage{
\method{summary}{mrcare}(object, ..., ci_level = 0.95, digits = 3)
}
\arguments{
\item{object}{Results object from mr_care function}

\item{ci_level}{Confidence level for intervals. Default is 0.95}

\item{digits}{Number of digits to round results to. Default is 3}
}
\value{
A list of class "summary.mrcare" containing formatted summary information
}
\description{
Provides a comprehensive summary of the results from an MRcare analysis,
including causal estimates, confidence intervals, and diagnostic information.
}
\examples{
\dontrun{
results <- mr_care(exposure_data, outcome_data)
summary(results)
}
}
