% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrcare-configuration.r
\name{test_configuration}
\alias{test_configuration}
\title{Test Configuration}
\usage{
test_configuration(config, verbose = TRUE)
}
\arguments{
\item{config}{Configuration list}

\item{verbose}{Whether to print detailed messages}
}
\value{
Logical indicating if configuration is working
}
\description{
Tests if the configuration is working properly
}
\keyword{internal}
