% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrcare-configuration.r
\name{test_plink}
\alias{test_plink}
\title{Test if PLINK is Working}
\usage{
test_plink(plink_path)
}
\arguments{
\item{plink_path}{Path to plink executable}
}
\value{
Logical indicating if plink works
}
\description{
Tests if a given plink executable is working
}
\keyword{internal}
